+++ src/renderer/utils/notifications/filters/product.test.ts
@@ -2,7 +2,7 @@ import {
   mockAccountNotifications,
   mockSingleAtlassifyNotification,
 } from '../../../__mocks__/notifications-mocks';
-import { defaultSettings } from '../../../context/App';
+import { defaultSettings } from '../../../context/defaults';
 import type { AtlassifyNotification, SettingsState } from '../../../types';
 import { PRODUCTS } from '../../products';
 import { productFilter } from '.';
